+++ backend/src/main/java/com/f12/moitz/application/dto/PlaceResponse.java
@@ -1,4 +1,4 @@
-package com.f12.moitz.ui.dto;
+package com.f12.moitz.application.dto;
 
 public record PlaceResponse(
         double latitude,
